+++ lib/generators/recording_studio_support/install/templates/recording_studio_support_initializer.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 RecordingStudioSupport.configure do |config|
-  # Support pages do not need extra config in this slice.
+  config.pages_path = "/support"
 end