+++ lib/recording_studio_support.rb
@@ -6,9 +6,12 @@ require "recording_studio_accessible"
 require "recording_studio_attachable"
 require "recording_studio_trashable"
 require "recording_studio_orderable"
+require "recording_studio_admin"
 require "recording_studio_support/version"
 require "recording_studio_support/engine"
 require "recording_studio_support/configuration"
+require "recording_studio_support/pages"
+require "recording_studio_support/admin"
 
 module RecordingStudioSupport
   class << self