+++ test/dummy/config/initializers/recording_studio_support.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 RecordingStudioSupport.configure do |config|
-  # Hosts can set options here. Support pages do not need extra config in this slice.
+  config.pages_path = "/support"
 end