+++ test/engine_test.rb
@@ -142,6 +142,26 @@ class EngineTest < Minitest::Test
     assert_equal 2, to_prepare_blocks.size
   end
 
+  def test_admin_initializer_registers_support_section
+    to_prepare_blocks = []
+    config_stub = Object.new
+    config_stub.define_singleton_method(:to_prepare) do |&block|
+      to_prepare_blocks << block
+    end
+
+    RecordingStudioSupport::Engine.stub(:config, config_stub) do
+      find_initializer("recording_studio_support.admin").block.call
+    end
+
+    registered = false
+    RecordingStudioSupport::Admin.stub(:register!, -> { registered = true }) do
+      to_prepare_blocks.first.call
+    end
+
+    assert_equal 1, to_prepare_blocks.size
+    assert registered
+  end
+
   def test_model_extension_initializer_skips_abstract_models
     to_prepare_blocks = []
     config_stub = Object.new